chore: rename package election_store to election_client
