+++ op-e2e/actions/helpers/l2_sequencer.go
@@ -18,7 +18,7 @@ import (
 	"github.com/ethereum-optimism/optimism/op-node/rollup/derive"
 	"github.com/ethereum-optimism/optimism/op-node/rollup/driver"
 	"github.com/ethereum-optimism/optimism/op-node/rollup/election"
-	"github.com/ethereum-optimism/optimism/op-node/rollup/election_store"
+	"github.com/ethereum-optimism/optimism/op-node/rollup/election_client"
 	"github.com/ethereum-optimism/optimism/op-node/rollup/engine"
 	"github.com/ethereum-optimism/optimism/op-node/rollup/event"
 	"github.com/ethereum-optimism/optimism/op-node/rollup/interop"
@@ -56,14 +56,14 @@ func NewL2Sequencer(t Testing, log log.Logger, l1 derive.L1Fetcher, blobSrc deri
 	beaconClient election.BeaconClient, altDASrc driver.AltDAIface, eng L2API, l1Client L1API, cfg *rollup.Config, seqConfDepth uint64,
 	interopBackend interop.InteropBackend) *L2Sequencer {
 
-	electionStore := election_store.NewElectionStore(log)
+	electionStore := election_client.NewElectionStore(log)
 	ver := NewL2Verifier(t, log, l1, blobSrc, beaconClient, altDASrc, eng, l1Client, cfg, &sync.Config{}, safedb.Disabled, interopBackend, electionStore)
 	attrBuilder := derive.NewFetchingAttributesBuilder(cfg, l1, eng)
 	seqConfDepthL1 := confdepth.NewConfDepth(seqConfDepth, ver.syncStatus.L1Head, l1)
 	l1OriginSelector := &MockL1OriginSelector{
 		actual: sequencing.NewL1OriginSelector(log, cfg, seqConfDepthL1),
 	}
-	electionClient := election_store.NewElectionClient(electionStore)
+	electionClient := election_client.NewElectionClient(electionStore)
 	metr := metrics.NoopMetrics
 	seqStateListener := node.DisabledConfigPersistence{}
 	conduc := &conductor.NoOpConductor{}
