+++ op-e2e/actions/sync/sync_test.go
@@ -25,7 +25,7 @@ import (
 	"github.com/ethereum-optimism/optimism/op-e2e/e2eutils"
 	"github.com/ethereum-optimism/optimism/op-node/rollup"
 	"github.com/ethereum-optimism/optimism/op-node/rollup/derive"
-	"github.com/ethereum-optimism/optimism/op-node/rollup/election_store"
+	"github.com/ethereum-optimism/optimism/op-node/rollup/election_client"
 	engine2 "github.com/ethereum-optimism/optimism/op-node/rollup/engine"
 	"github.com/ethereum-optimism/optimism/op-node/rollup/event"
 	"github.com/ethereum-optimism/optimism/op-node/rollup/sync"
@@ -830,7 +830,7 @@ func TestELSyncTransitionsToCLSyncAfterNodeRestart(gt *testing.T) {
 
 	PrepareELSyncedNode(t, miner, sequencer, seqEng, verifier, verEng, seqEngCl, batcher, dp)
 
-	electionStore := election_store.NewElectionStore(logger)
+	electionStore := election_client.NewElectionStore(logger)
 	// Create a new verifier which is essentially a new op-node with the sync mode of ELSync and default geth engine kind.
 	verifier = actionsHelpers.NewL2Verifier(t, captureLog, l1Cl, miner.BlobStore(), miner.BeaconClient(), altda.Disabled, verifier.Eng, l1Cl.EthClient, sd.RollupCfg, &sync.Config{SyncMode: sync.ELSync}, actionsHelpers.DefaultVerifierCfg().SafeHeadListener, nil, electionStore)
 
@@ -875,7 +875,7 @@ func TestForcedELSyncCLAfterNodeRestart(gt *testing.T) {
 
 	PrepareELSyncedNode(t, miner, sequencer, seqEng, verifier, verEng, seqEngCl, batcher, dp)
 
-	electionStore := election_store.NewElectionStore(logger)
+	electionStore := election_client.NewElectionStore(logger)
 	// Create a new verifier which is essentially a new op-node with the sync mode of ELSync and erigon engine kind.
 	verifier2 := actionsHelpers.NewL2Verifier(t, captureLog, l1Cl, miner.BlobStore(), miner.BeaconClient(), altda.Disabled, verifier.Eng, l1Cl.EthClient, sd.RollupCfg, &sync.Config{SyncMode: sync.ELSync, SupportsPostFinalizationELSync: true}, actionsHelpers.DefaultVerifierCfg().SafeHeadListener, nil, electionStore)
